move 'move' tests to separate file
