+++ types/ramda/tsconfig.json
@@ -147,6 +147,7 @@
         "test/mergeWithKey-tests.ts",
         "test/min-tests.ts",
         "test/minBy-tests.ts",
-        "test/modulo-tests.ts"
+        "test/modulo-tests.ts",
+        "test/move-tests.ts"
     ]
 }
